Fixed conversion from and to Decimal
Update version and changelog

Update usage of Decimal in JDBC

Update Decimal tests

Added scale-based methods for init DecimalValue, old raw-based marked as deprecated

Fixed JDBC tests by removing the BigDecimal 'adjustments'

Adjust for input scale in DecimalValue newValue(BigDecimal)

+++ jdbc/src/main/java/com/yandex/ydb/jdbc/impl/MappingGetters.java
@@ -301,7 +301,7 @@ public class MappingGetters {
                     };
             }
         } else if (kind == Type.Kind.DECIMAL) {
-            return value -> value.getDecimal().toBigInteger().longValue(); // TODO: Improve performance
+            return value -> value.getDecimal().toLong();
         } else {
             return value -> {
                 throw dataTypeNotSupported(kind, javaType);
